Merge branch 'development' into ssh-config

+++ src/Command/Project/ProjectDeleteCommand.php
@@ -26,8 +26,8 @@ class ProjectDeleteCommand extends CommandBase
             return 1;
         }
 
-        /** @var \Platformsh\Cli\Helper\QuestionHelper $questionHelper */
-        $questionHelper = $this->getHelper('question');
+        /** @var \Platformsh\Cli\Service\QuestionHelper $questionHelper */
+        $questionHelper = $this->getService('question_helper');
 
         $confirmQuestionLines = [
             'You are about to delete the project:',
@@ -55,6 +55,9 @@ class ProjectDeleteCommand extends CommandBase
 
         $subscriptionId = $project->getSubscriptionId();
         $subscription = $this->api()->getClient()->getSubscription($subscriptionId);
+        if (!$subscription) {
+            throw new \RuntimeException('Subscription not found: ' . $subscriptionId);
+        }
 
         $subscription->delete();
         $this->api()->clearProjectsCache();
